배정 클랜원 표기 추가 및 이동 필요 필터링 기능 추가
