+++ src/main/java/open/api/coc/external/coc/clan/ClanApiServiceImpl.java
@@ -94,10 +94,11 @@ public class ClanApiServiceImpl implements ClanApiService {
     }
 
     private Optional<Player> findPlayer(String playTag) {
+        log.info("{} [{}]", clashOfClanConfig.getPlayerUri(), playTag);
         return Optional.ofNullable(restClient.get()
-                .uri(clashOfClanConfig.getPlayerUri(), playTag)
-                .retrieve()
-                .body(Player.class));
+                                             .uri(clashOfClanConfig.getPlayerUri(), playTag)
+                                             .retrieve()
+                                             .body(Player.class));
     }
 
     @Override
