+++ src/main/resources/static/css/common.css
@@ -193,7 +193,7 @@ body::before {
   font-weight: bold;
   word-break: break-all;
   border: var(--default-border);
-  padding: 3px 25px;
+  padding: 3px 15px;
   border-radius: 15px;
   background: #5c5c5c;
   color: white;
