+++ src/main/resources/templates/capital/Raid.html
@@ -11,6 +11,7 @@
   <link rel="stylesheet" type="text/css" href="/static/css/clan/raid.css" />
 
   <script src="/static/js/common.js"></script>
+  <script src="/static/js/util.js"></script>
   <script src="/static/js/api/league.js"></script>
   <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/dayjs/1.11.10/dayjs.min.js" integrity="sha512-FwNWaxyfy2XlEINoSnZh1JQ5TRRtGow0D6XcmAWmYCRgvqOUTnzCxPc9uF35u5ZEpirk1uhlPVA19tflhvnW1g==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
