+++ src/main/resources/templates/cms/ClanManager.html
@@ -12,6 +12,7 @@
   <link rel="stylesheet" type="text/css" href="/static/css/cms/manager.css" />
 
   <script src="/static/js/common.js"></script>
+  <script src="/static/js/util.js"></script>
   <script src="/static/js/api/league.js"></script>
   <script src="/static/js/api/clan.js"></script>
